+++ main/lib/src/FileWriterStandard.cc
@@ -11,7 +11,7 @@ namespace eudaq {
       FileWriterStandard(const std::string &);
       virtual void StartRun(unsigned);
       virtual void WriteEvent(const DetectorEvent &);
-      virtual unsigned long long FileBytes() const;
+      virtual uint64_t FileBytes() const;
       virtual ~FileWriterStandard();
     private:
       FileSerializer * m_ser;
@@ -41,6 +41,6 @@ namespace eudaq {
     delete m_ser;
   }
 
-  unsigned long long FileWriterStandard::FileBytes() const { return m_ser ? m_ser->FileBytes() : 0; }
+  uint64_t FileWriterStandard::FileBytes() const { return m_ser ? m_ser->FileBytes() : 0; }
 
 }
